@@ -1,18 +1,14 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { SRCSET_ATTRS } from '@angular/core/src/sanitization/html_sanitizer';
+import { Component, OnInit, Input } from "@angular/core";
 
 @Component({
-  selector: 'app-blog-item',
-  templateUrl: './blog-item.component.html',
-  styleUrls: ['./blog-item.component.scss']
+  selector: "app-blog-item",
+  templateUrl: "./blog-item.component.html",
+  styleUrls: ["./blog-item.component.scss"]
 })
 export class BlogItemComponent implements OnInit {
-
   @Input() src;
 
-  constructor() { }
+  constructor() {}
 
-  ngOnInit() {
-  }
-
+  ngOnInit() {}
 }
