@@ -55,13 +55,7 @@ import "web-animations-js"; // Run `npm install --save web-animations-js`.
 /***************************************************************************************************
  * Zone JS is required by default for Angular itself.
  */
-import "zone.js/dist/zone"; // Included with Angular CLI.
-
-/***************************************************************************************************
- * APPLICATION IMPORTS
- */
-
-import "core-js/es6/symbol";
+import "zone.js/dist/zone";
 import "core-js/es6/object";
 import "core-js/es6/function";
 import "core-js/es6/parse-int";
